fix(dashboard): improve config schema UX — ranges, selects, i18n labels
- Add min/max/step to all number fields across every config section
- Change fixed-value fields to select (language, embedding_provider,
  image_provider) with { value, label } i18n support in ConfigPage
- Introduce number_select type for interval fields (consolidation,
  cleanup) with translated preset labels via i18n keys
- Fix decay_rate f32→f64 to eliminate floating-point display artifacts
- Show workspaces_dir effective path instead of null when unset
- Sync registry provider files on update, not just on first copy
- Show generating placeholder in ChatInput while response is loading
- Add missing desc_*/fld_* i18n keys for all new constrained fields

+++ crates/librefang-kernel/src/kernel/mod.rs
@@ -1350,7 +1350,7 @@ impl LibreFangKernel {
             .unwrap_or_else(|| config.data_dir.join("librefang.db"));
         let mut substrate = MemorySubstrate::open_with_chunking(
             &db_path,
-            config.memory.decay_rate,
+            config.memory.decay_rate as f32,
             config.memory.chunking.clone(),
         )
         .map_err(|e| KernelError::BootFailed(format!("Memory init failed: {e}")))?;
